Fixes to network.sh for dnsmasq and missing eth1 interface
Puppet starts the dnsmasq service. The service needs to stopped
to allow the default network to come up. The dnsmasq service will
start back up with the default network.

eth1 is set as the public_interface in the local heat metadata.
Not all environments will have an eth1. Added some logic to
change eth1 to eth0 if eth0 is the interface that is present.