Don't fail if firewalld is not installed
firewalld is not installed on the Fedora cloud images, which means
it can't be stopped or masked during the step that makes the
system use iptables.  Print a message instead of failing the
script in that situation.

(cherry picked from commit 6bf6c9d1c2c31e3130cc93e5687a691706856319)